chore(gateway-router)!: renamed the gaewayApolloClient to eventsFederationApolloClient
LIIKUNTA-320
