+++ apps/events-helsinki/src/hooks/useRHHCConfig.tsx
@@ -10,7 +10,7 @@ import Link from 'next/link';
 import React from 'react';
 import type { Config } from 'react-helsinki-headless-cms';
 import { defaultConfig as rhhcDefaultConfig } from 'react-helsinki-headless-cms';
-import { useApolloClient } from 'domain/clients/gatewayApolloClient';
+import { useApolloClient } from 'domain/clients/eventsFederationApolloClient';
 
 import AppConfig from '../domain/app/AppConfig';
 import type { ArticleDetailsProps } from '../domain/article/articleDetails/ArticleDetails';
