+++ apps/hobbies-helsinki/src/pages/articles/[...slug].tsx
@@ -26,7 +26,7 @@ import type {
   ArticleQueryVariables,
 } from 'react-helsinki-headless-cms/apollo';
 import { ArticleDocument } from 'react-helsinki-headless-cms/apollo';
-import { createApolloClient } from 'domain/clients/gatewayApolloClient';
+import { createApolloClient } from 'domain/clients/eventsFederationApolloClient';
 import Navigation from '../../common-events/components/navigation/Navigation';
 import ShareLinks from '../../common-events/components/shareLinks/ShareLinks';
 import {
